polling clean up
